+++ R/waiter.R
@@ -44,7 +44,7 @@
 #'   
 #'   observeEvent(input$show, {
 #'     waiter_show(
-#'       tagList(
+#'       html = tagList(
 #'         spin_fading_circles(),
 #'         "Loading ..."
 #'       )