+++ NEWS.md
@@ -7,6 +7,7 @@
 - Added a convenience function `transparent` to easily create transparent waiter backgrounds.
 - Added a cheat sheet, thanks to [JH Kim](https://github.com/jhk0530)
 - Add ability to reduce load size of CSS (raised by issue [#47](https://github.com/JohnCoene/waiter/issues/47)), `use_waiter` takes a `spinner` argument to which one can specify any of 7 spinner CSS kits, by default all kits are loaded. One can know which kits should be specified by simply typing the spinner in the console, e.g.: `spin_rotating_plane()`.
+- Fix waiter example.
 
 ## Bug Fixes